# docs/api-openapi-draft.yaml
openapi: 3.1.0
info:
  title: Nexus Arcade Hub and Player Passport API
  version: 0.1.0
  description: Pilot V1 API for online cabinet QR pairing, player profiles, game sessions, progression, and leaderboards. QR pairing URLs carry only a cabinet session ID; player identity proof comes from the phone-authenticated request.
servers:
  - url: http://localhost:3000
paths:
...
  /api/arcade/cabinet-login/{sessionId}/status:
    get:
      summary: Poll QR pairing status
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Pairing status
...
  /api/operator/config:
    get:
      summary: Return protected pilot runtime configuration
      security:
        - OperatorSession: []
      responses:
        "200":
          description: Operator configuration
    patch:
      summary: Update protected pilot runtime configuration
      security:
        - OperatorSession: []
      responses:
        "200":
          description: Updated operator configuration
...
  /api/player/claim-cabinet-session:
    post:
      summary: Claim an online cabinet QR session for a phone-authenticated Player Passport
      security:
        - PlayerSession: []
      responses:
        "200":
          description: Player assigned to cabinet slot
...
components:
  securitySchemes:
    OperatorSession:
      type: apiKey
      in: header
      name: x-operator-token
    PlayerSession:
      type: apiKey
      in: header
      name: x-player-id
    PassportBearer:
      type: http
      scheme: bearer